Modification after testing Linux system
